fix error code

+++ Kindom/Assets/Script/Common/CG/CGConvex.cs
@@ -21,7 +21,7 @@ namespace Common.CG
 
 			int count = points.Length;
 
-			int lastDirection = -1;
+			float lastDirection = -1;
 
 			for (int i = 0; i < count; i++)
 			{
@@ -32,7 +32,7 @@ namespace Common.CG
 				Vector2 v0 = p1 - p0;
 				Vector2 v1 = p2 - p1;
 
-				int vd = Vector2.Dot (v0, v1);
+				float vd = Vector2.Dot (v0, v1);
 				vd = vd > 0 ? 1 : vd == 0 ? 0 : 2;
 				if (vd == 0)
 				{
